Back-end/src/main/java/com/example/webapp/services/UserDetailsServiceImpl.java
package com.example.webapp.services;

import com.example.webapp.models.User;
import com.example.webapp.security.UserDetailsImpl;
import com.example.webapp.repositories.UserRepository;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

import javax.transaction.Transactional;

@Service
public class UserDetailsServiceImpl implements UserDetailsService {

    //Importing User repository
    private final UserRepository userRepository;

    public UserDetailsServiceImpl(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    //Finds user by username in the database
    @Override
    @Transactional
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        User user = userRepository.findByUsername(username)
                .orElseThrow(() -> new UsernameNotFoundException("User with username: [" + username +  "] Not Found"));

        //Building the UserDetails object using the retrieved user object from the database
        return UserDetailsImpl.build(user);
    }
}
